+++ web/modules/custom/hoeringsportal_citizen_proposal/src/Form/ProposalFormAdd.php
@@ -115,14 +115,16 @@ final class ProposalFormAdd extends ProposalFormBase {
     ];
 
     $form['proposal'] = [
-      '#type' => 'textarea',
+      '#type' => 'text_format',
+      '#format' => self::CONTENT_TEXT_FORMAT,
+      '#allowed_formats' => [self::CONTENT_TEXT_FORMAT],
       '#required' => TRUE,
       '#rows' => 15,
       '#title' => $this
         ->t('Proposal'),
       '#description' => $this->getAdminFormStateValue('proposal_help'),
       '#description_display' => 'before',
-      '#default_value' => $defaltValues['proposal'],
+      '#default_value' => $defaltValues['proposal'] ?? '',
       '#maxlength_js' => TRUE,
       '#attributes' => [
         'data-maxlength' => $this->getMaxLength('characters_proposal'),
@@ -131,14 +133,16 @@ final class ProposalFormAdd extends ProposalFormBase {
     ];
 
     $form['remarks'] = [
-      '#type' => 'textarea',
+      '#type' => 'text_format',
+      '#format' => self::CONTENT_TEXT_FORMAT,
+      '#allowed_formats' => [self::CONTENT_TEXT_FORMAT],
       '#required' => TRUE,
       '#rows' => 15,
       '#title' => $this
         ->t('Remarks'),
       '#description' => $this->getAdminFormStateValue('remarks_help'),
       '#description_display' => 'before',
-      '#default_value' => $defaltValues['remarks'],
+      '#default_value' => $defaltValues['remarks'] ?? '',
       '#maxlength_js' => TRUE,
       '#attributes' => [
         'data-maxlength' => $this->getMaxLength('characters_remarks'),
@@ -165,6 +169,34 @@ final class ProposalFormAdd extends ProposalFormBase {
         : $this->t('Create proposal'),
       '#button_type' => 'primary',
     ];
+
+    $form['#after_build'][] = $this->afterBuildForm(...);
+
+    return $form;
+  }
+
+  /**
+   * Form after build handler.
+   *
+   * @param array $form
+   *   The form.
+   * @param \Drupal\Core\Form\FormStateInterface $formState
+   *   The form state.
+   *
+   * @return array
+   *   The form.
+   */
+  private function afterBuildForm(array $form, FormStateInterface $formState) {
+    // Hide text format info
+    // (lifted from _allowed_formats_remove_textarea_help()).
+    foreach ($form as $key => &$element) {
+      if (isset($element['format'])) {
+        unset(
+          $element['format']['help'],
+          $element['format']['guidelines']
+        );
+      }
+    }
 
     return $form;
   }
@@ -216,11 +248,20 @@ final class ProposalFormAdd extends ProposalFormBase {
       $form_state->setErrorByName('title', $this->t('Too many characters used.'));
     }
 
-    if (strlen($form_state->getValue('proposal')) > $this->getMaxLength('characters_proposal')) {
+    $getTextValue = function (string $key) use ($form_state) {
+      $value = $form_state->getValue($key);
+      if (isset($value['value'])) {
+        $value = $value['value'];
+      }
+
+      return strip_tags((string) $value);
+    };
+
+    if (strlen($getTextValue('proposal')) > $this->getMaxLength('characters_proposal')) {
       $form_state->setErrorByName('proposal', $this->t('Too many characters used.'));
     }
 
-    if (strlen($form_state->getValue('remarks')) > $this->getMaxLength('characters_remarks')) {
+    if (strlen($getTextValue('remarks')) > $this->getMaxLength('characters_remarks')) {
       $form_state->setErrorByName('remarks', $this->t('Too many characters used.'));
     }
   }
@@ -237,14 +278,8 @@ final class ProposalFormAdd extends ProposalFormBase {
       'field_author_phone' => $formState->getValue('phone'),
       'field_author_email' => $formState->getValue('email'),
       'field_author_email_display' => $formState->getValue('email_display'),
-      'field_proposal' => [
-        'value' => $formState->getValue('proposal'),
-        'format' => 'filtered_html',
-      ],
-      'field_remarks' => [
-        'value' => $formState->getValue('remarks'),
-        'format' => 'filtered_html',
-      ],
+      'field_proposal' => $formState->getValue('proposal'),
+      'field_remarks' => $formState->getValue('remarks'),
     ]);
     $this->helper->setDraftProposal($entity);
     $formState