+++ CHANGELOG.md
@@ -6,6 +6,8 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+* [PR-351](https://github.com/itk-dev/hoeringsportal/pull/351)
+  Added and used citizen proposal text format
 * [PR-347](https://github.com/itk-dev/hoeringsportal/pull/347)
   Added citizen proposal survey
 * [PR-342](https://github.com/itk-dev/hoeringsportal/pull/342)